Doing initial avif output

+++ src/main.rs
@@ -11,7 +11,7 @@ fn main() -> anyhow::Result<()> {
         PathBuf::from("/Users/alan/Documents/Neopoligen/alanwsmith.com/avif_test_images");
     let output_root = PathBuf::from("/Users/alan/Documents/Neopoligen/alanwsmith.com/cache/images");
     let extensions = vec!["jpg", "png", "jpeg"];
-    let max_widths = vec![100, 200, 300];
+    let max_widths = vec![100, 200, 300, 400];
     let source_files = get_files_with_extensions(&input_dir, &extensions);
     source_files.iter().for_each(|f| {
         let output_base_dir = output_root.join(f.file_stem().unwrap());
@@ -54,8 +54,6 @@ fn make_avif(
     output_base_dir: &PathBuf,
     max_width: u32,
 ) -> anyhow::Result<()> {
-    //match output_path.parent() {
-    //   Some(parent_dir) =>
     match fs::create_dir_all(output_base_dir) {
         Ok(_) => {
             let img_file = ImageReader::open(input_path)?.decode()?;
@@ -67,30 +65,32 @@ fn make_avif(
             let output_height = img_file.height() * output_width / img_file.width();
             let output_path = output_base_dir.join(format!("{}w.avif", output_width));
 
-            println!("Making: {}", output_path.display());
+            if !file_exists(&output_path) {
+                println!("Making: {}", output_path.display());
 
-            let resized_image = img_file.resize_to_fill(
-                output_width,
-                output_height,
-                image::imageops::FilterType::Lanczos3,
-            );
-
-            let img = Img::new(
-                resized_image.as_bytes().as_rgb(),
-                resized_image.width() as usize,
-                resized_image.height() as usize,
-            );
-            let res = Encoder::new()
-                .with_quality(70.)
-                .with_speed(4)
-                .encode_rgb(img)?;
-            std::fs::write(output_path, res.avif_file)?;
+                let resized_image = img_file.resize_to_fill(
+                    output_width,
+                    output_height,
+                    image::imageops::FilterType::Lanczos3,
+                );
+                let img = Img::new(
+                    resized_image.as_bytes().as_rgb(),
+                    resized_image.width() as usize,
+                    resized_image.height() as usize,
+                );
+                let res = Encoder::new()
+                    .with_quality(70.)
+                    .with_speed(4)
+                    .encode_rgb(img)?;
+                std::fs::write(output_path, res.avif_file)?;
+            } else {
+                println!("Exists: {}", output_path.display());
+            }
         }
         Err(e) => {
             dbg!(e);
             ()
-        } // },
-          // None => println!("something went wrong"),
+        }
     }
 
     // let img_file = ImageReader::open(input_path)?.decode()?;
@@ -106,4 +106,17 @@ fn make_avif(
     // std::fs::write(output_path, res.avif_file)?;
 
     Ok(())
+}
+
+fn file_exists(path: &PathBuf) -> bool {
+    match path.try_exists() {
+        Ok(exists) => {
+            if exists == true {
+                true
+            } else {
+                false
+            }
+        }
+        Err(_) => false,
+    }
 }
